feat: abstract task logic to core
